+++ CHANGELOG.md
@@ -33,7 +33,7 @@ and this project adheres to
 
 ### Changed
 
-- Upgraded `@jupiterone/integration-sdk-*@6.15.0`
+- Upgraded `@jupiterone/integration-sdk-*@6.18.0`
 
 ### Remvoved